+++ beaker/decorators_test.py
@@ -4,6 +4,7 @@ import pyteal as pt
 from .decorators import (
     external,
     get_handler_config,
+    ResolvableArgument,
     Authorize,
     create,
     clear_state,
@@ -254,3 +255,41 @@ def test_bare():
 
     hc = get_handler_config(impl)
     assert hc.bare_method.clear_state.action.subroutine.implementation == impl
+
+
+def test_resolvable():
+    from .state import (
+        AccountStateValue,
+        ApplicationStateValue,
+        DynamicAccountStateValue,
+        DynamicApplicationStateValue,
+    )
+
+    x = AccountStateValue(pt.TealType.uint64, key=pt.Bytes("x"))
+    r = ResolvableArgument(x)
+    assert r.resolvable_class == "local-state"
+
+    x = DynamicAccountStateValue(pt.TealType.uint64, max_keys=1)
+    r = ResolvableArgument(x[pt.Bytes("x")])
+    assert r.resolvable_class == "local-state"
+
+    x = ApplicationStateValue(pt.TealType.uint64, key=pt.Bytes("x"))
+    r = ResolvableArgument(x)
+    assert r.resolvable_class == "global-state"
+
+    x = DynamicApplicationStateValue(pt.TealType.uint64, max_keys=1)
+    r = ResolvableArgument(x[pt.Bytes("x")])
+    assert r.resolvable_class == "global-state"
+
+    @external(read_only=True)
+    def x():
+        return pt.Assert(pt.Int(1))
+
+    r = ResolvableArgument(x)
+    assert r.resolvable_class == "abi-method"
+
+    r = ResolvableArgument(pt.Bytes("1"))
+    assert r.resolvable_class == "constant"
+
+    r = ResolvableArgument(pt.Int(1))
+    assert r.resolvable_class == "constant"
